services/story-service/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.4
django-redis==5.4.0
kafka-python==2.0.2
celery==5.4.0
dj-database-url==2.1.0
gunicorn==22.0.0
django-prometheus==2.3.1
